+++ packages/patrol_cli/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Add `--web-server-timeout` flag to configure the web server startup timeout (default: 120s). (#2948)
 - Fix running devtools after moving test to `patrol_test/`. (#2922) 
 - Fix listen to stdout/stderr streams in buildApkConfigOnly to prevent hang on Windows. (#2905)
 - Fix bug that --exclude was not working for directories. (#2918)